feat(#88): refactor: format-on-save.ts — audit and modularize
