+++ .pi/extensions/format-on-save/types.mts
@@ -1,0 +1,9 @@
+/** ESLint diagnostic entry parsed from JSON output. */
+export interface EslintDiagnostic {
+	file: string;
+	line: number;
+	column: number;
+	severity: "Error" | "Warning";
+	message: string;
+	ruleId: string | null;
+}
